+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-08-20T18:41:40.524Z",
+    "2026-08-21T08:04:33.837Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {